first pass mt solver, not reporting right results
