+++ src/solver.ts
@@ -1,44 +1,35 @@
-import { onLetterPressed } from "./letters";
-import { updateState } from "./gameState";
-import { LetterEntity, Letter } from "./letterEntity";
-import { Gameboard, maxY, maxX } from "./board";
-import { TypedPriorityQueue } from "../libs/TypedPriorityQueue";
 import { getLevel } from "./levels";
-import { Strategy, strategyList } from "./strategy";
+import { strategyList } from "./strategy";
+import { Solution } from "./solverWorker";
+import { Gameboard } from "./board";
 
-const BudgetPerRunMS = 100;
+export async function solveMany(start: number, end: number) {
+    const begin = performance.now();
+    //Init worker queue
+    const workerQueue = new Array<Worker>();
+    for (let i = 0; i < 1; ++i) {
+        workerQueue.push(new Worker('solverWorker.js'));
+    }
 
-type Move = { x: number, y: number };
-type Path = { moves: Array<Move>, score: number, state: Gameboard };
-
-interface Run {
-    strategy: string,
-    steps: number,
-    numberOfSolutions: number,
-    firstSolutionTime: number,
-    bestPath: Path,
-    bestScore: number,
-    shortestPathLength: number
-}
-
-interface Solution {
-    solved: boolean
-    runs: Run[]
-    shortestPathLength: number,
-    bestPath: Path,
-}
-
-
-export function solveMany(start: number, end: number) {
-    const data = new Array<Solution>();
+    const working = new Array<Promise<Solution>>();
     for (let i = start; i < end; ++i) {
         const level = getLevel(i);
-        const solution = solve(level);
-        data.push(solution);
-        console.count("Processed");
+        working.push(solveWithWorkerQueue(level, workerQueue));
     }
+    const data = await Promise.all(working);
+
+    const finish = performance.now();
+    data.sort((a, b) => b.level - a.level);
+    console.log(`Finished, at: ${finish - begin}`);
+
     const csv = formatCSV(data, start);
     download('levels.csv', csv);
+}
+
+export async function solve(gameboard: Gameboard) {
+    const worker = new Worker('solverWorker.js');
+    const solution = await solveWithWorkerQueue(gameboard, [worker]);
+    return solution;
 }
 
 export function printSolution(solution: Solution) {
@@ -56,174 +47,38 @@ export function printSolution(solution: Solution) {
     }
 }
 
-export function solve(board: Gameboard): Solution {
+const workRequests = new Array<(worker: Worker) => void>();
 
-    let pathQueue: TypedPriorityQueue<Path>;
+async function solveWithWorkerQueue(level: Gameboard, workerQueue: Array<Worker>) {
 
-    let solved = false;
-    let currentRun = 0;
-    const runs = new Array<Run>();
-
-    do {
-        //start new run
-        pathQueue = new TypedPriorityQueue<Path>(comparePathByScore);
-        pathQueue.add({ moves: [], score: 0, state: board });
-        const runWins = new Array<Path>();
-        const runStart = performance.now();
-        let runTime = 0;
-        let steps = 0;
-        let firstSolutionTime = 0;
-        let runSolved = false;
-
-        while (runTime < BudgetPerRunMS && !pathQueue.isEmpty()) {
-            step(pathQueue.poll()!, strategyList[currentRun], runWins);
-            ++steps;
-
-            const end = performance.now();
-            runTime = end - runStart;
-
-            if (!runSolved && runWins.length > 0) {
-                solved = true;
-                runSolved = true;
-                firstSolutionTime = runTime;
-            }
-        }
-
-        //Post-run
-        const numberOfSolutions = runWins.length;
-        let shortestPathLength;
-        let bestPath: Path;
-        if (runSolved) {
-            bestPath = (runWins.length === 1)
-                ? runWins.pop()!
-                : runWins.reduce((a, b) => a.moves.length < b.moves.length ? a : b);
-            shortestPathLength = bestPath.moves.length;
+    function onWorkerDone(worker: Worker) {
+        if (workRequests.length > 0) {
+            const req = workRequests.pop()!
+            req(worker);
         }
         else {
-            shortestPathLength = -1;
-            bestPath = pathQueue.peek()!;
+            workerQueue.push(worker);
         }
-        const bestScore = bestPath.score;
+    }
 
-        runs.push({
-            strategy: strategyList[currentRun].name,
-            steps, numberOfSolutions, firstSolutionTime,
-            bestPath, shortestPathLength, bestScore
+    async function getWorker() {
+        if (workerQueue.length > 0) {
+            return workerQueue.pop()!;
+        }
+        else return new Promise<Worker>((resolve, reject) => {
+            workRequests.push(worker => resolve(worker));
         });
-        currentRun++;
-
-    } while (currentRun < strategyList.length && !pathQueue.isEmpty())
-
-    const solvedRuns = runs.filter(run => run.numberOfSolutions > 0);
-    let bestRun: Run;
-    if (solvedRuns.length > 0) {
-        bestRun = solvedRuns.reduce((a, b) => a.shortestPathLength < b.shortestPathLength ? a : b);
     }
-    else {
-        bestRun = runs.reduce((a, b) => a.bestScore > b.bestScore ? a : b);
-    }
-
-    return {
-        solved,
-        runs,
-        shortestPathLength: bestRun.shortestPathLength,
-        bestPath: bestRun.bestPath
-    };
-
-    function step(path: Path, weights: Strategy, wins: Array<Path>) {
-        const availableMoves = getMoves(path.state);
-        if (availableMoves.length === 0) { return }
-
-        for (let i = 0; i < availableMoves.length; ++i) {
-            let testBoard = path.state.clone();
-            let testBoardMoves = getMoves(testBoard);
-            const letter = testBoardMoves[i];
-            testBoard = doMove(testBoard, letter);
-            let score = evaluate(testBoard, path.moves.length, weights);
-            if (score === -Infinity) continue; //Drop losing paths 
-            let move = { x: letter.x, y: letter.y };
-            const newPath = {
-                moves: path.moves.concat(move),
-                score: score,
-                state: testBoard
-            };
-            if (score == Infinity) {
-                wins.push(newPath);
-            }
-            else {
-                pathQueue.add(newPath);
-            }
+    const worker = await getWorker();
+    return new Promise<Solution>((resolve, reject) => {
+        worker.onmessage = (message: any) => {
+            const solution = message.data as Solution;
+            console.count(`Processed ${level}`);
+            onWorkerDone(worker);
+            resolve(solution);
         }
-    }
-}
-
-
-function getMoves(board: Gameboard) {
-    return board.entities.filter(entity => canClick(entity.letter));
-}
-
-function doMove(board: Gameboard, clickedEntity: LetterEntity) {
-    let effects = onLetterPressed(board, clickedEntity);
-    while (effects.length !== 0) {
-        effects = updateState(board, effects);
-    }
-    return board; //transformed in updateState
-}
-
-function evaluate(board: Gameboard, moveCount: number, weights: Strategy): number {
-    const scored = (board.firstLetterScored ? 1 : 0) + (board.secondLetterScored ? 1 : 0) + (board.thirdLetterScored ? 1 : 0);
-    if (scored === 3) return Infinity; //Won
-
-    const first = board.entities.find(state => state.letter === Letter.First);
-    const second = board.entities.find(state => state.letter === Letter.Second);
-    const third = board.entities.find(state => state.letter === Letter.Third);
-
-    if ((!first && !board.firstLetterScored)
-        || (!second && !board.secondLetterScored)
-        || (!third && !board.thirdLetterScored)
-    ) return -Infinity; //Lost
-
-    const firstHeight = !first ? 0 : normalizedDistanceFromTop010(first.y);
-    const secondHeight = !second ? 0 : normalizedDistanceFromTop010(second.y);
-    const thirdHeight = !third ? 0 : normalizedDistanceFromTop010(third.y);
-
-    const invisibles = board.entities.filter(state => state.letter === Letter.I).length;
-    const invisiblesRemoved = (maxY * maxX) - invisibles;
-
-    const scoreComponent = weights.score * scored;
-    const heightComponent = weights.allLetter * (firstHeight + secondHeight + thirdHeight);
-    const firstHeightComponent = weights.firstLetter * firstHeight;
-    const secondHeightComponent = weights.secondLetter * secondHeight;
-    const thirdHeightComponent = weights.thirdLetter * thirdHeight;
-    const invisComponent = weights.invisibles * invisiblesRemoved;
-    const letterCountComponent = weights.letterCount * board.entities.length;
-
-    const pathLengthComponent = -0.01 * moveCount; //between two equal paths, the one with fewer steps is better
-
-    return heightComponent + firstHeightComponent + secondHeightComponent + thirdHeightComponent +
-        scoreComponent + invisComponent + pathLengthComponent + letterCountComponent;
-
-    //distance from top (y) normalized and multiplied to 1-10 range
-    function normalizedDistanceFromTop010(y: number) {
-        const bottom = (maxY - 1);
-        return (y / bottom) * 10
-    }
-}
-
-function canClick(letter: Letter) {
-    switch (letter) {
-        case Letter.I:
-        case Letter.First:
-        case Letter.Second:
-        case Letter.Third:
-            return false;
-        default:
-            return true;
-    }
-}
-
-function comparePathByScore(a: Path, b: Path) {
-    return a.score > b.score;
+        worker.postMessage(level);
+    });
 }
 
 function formatCSV(data: Array<Solution>, startLevel: number) {
